// wardn/frontend/lib/api/generated/model/mCPRuntimeEventRead.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPRuntimeEventReadMetadata } from './mCPRuntimeEventReadMetadata';

export interface MCPRuntimeEventRead {
  createdAt: string;
  eventType: string;
  id: string;
  message: string;
  metadata?: MCPRuntimeEventReadMetadata;
  runtimeSessionId: string;
}