// wardn/frontend/lib/api/generated/model/mCPRuntimeSessionHealthResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPRuntimeSessionHealthResponseDetails } from './mCPRuntimeSessionHealthResponseDetails';

export interface MCPRuntimeSessionHealthResponse {
  details?: MCPRuntimeSessionHealthResponseDetails;
  healthy: boolean;
  message: string;
  ready: boolean;
  runtimeKind: string;
  runtimeProvider: string;
  runtimeSessionId: string;
  status: string;
}